feat: Add new dataproc_magics module providing `%pip install` features
The initial goal of this module is to shadow the built-in `%pip` magic
to support installing from `gs://` URLs in addition to local files and
`https://` URLs. (Other magic implementations may be added later.)

In this commit, we support only the following constructions:

- %pip install gs://bucket/mypackage.whl
- %pip install -r gs://bucket/myrequirements.txt

Recursively handling `gs://` URLs inside requirements files is not yet
supported.